+++ clang/lib/Driver/ToolChains/CommonArgs.cpp
@@ -3334,7 +3334,7 @@ void tools::handleVectorizeSLPArgs(const ArgList &Args,
 
 void tools::handleInterchangeLoopsArgs(const ArgList &Args,
                                        ArgStringList &CmdArgs) {
-  // FIXME: Instead of relying on shouldEnableVectorizerAtOLevel, we may want to
+  // FIXME: instead of relying on shouldEnableVectorizerAtOLevel, we may want to
   // implement a separate function to infer loop interchange from opt level.
   // For now, enable loop-interchange at the same opt levels as loop-vectorize.
   bool EnableInterchange = shouldEnableVectorizerAtOLevel(Args, false);