+++ flang/docs/ReleaseNotes.md
@@ -35,9 +35,7 @@ page](https://llvm.org/releases/).
 
 ## New Compiler Flags
 
-* -floop-interchange is now recognized by flang.
-* -floop-interchange is enabled by default at -O2 and above.
-* -floop-fuse is now recognized by flang.
+* -fexperimental-loop-fusion is now recognized by flang.
 
 ## Windows Support